Delete enomorous classes. Work with chat
