+++ app/build.gradle
@@ -61,9 +61,6 @@ android {
 }
 
 dependencies {
-    implementation 'com.google.firebase:firebase-messaging-ktx:23.1.1'
-    implementation 'com.google.firebase:firebase-auth-ktx:21.1.0'
-    implementation 'com.google.firebase:firebase-database-ktx:20.1.0'
     def retrofit_version = '2.9.0'
     def gson_version = '2.10'
     def gson_converter_version = '2.9.0'
