--- app/src/main/java/com/example/roomer/domain/model/ChatMessage.kt
@@ -1,10 +1,0 @@
-package com.example.roomer.domain.model
-
-import java.util.Date
-
-data class ChatMessage(
-    var messageText: String,
-    var messageSenderUser: String,
-    var messageReceiverUser: String,
-    var messageTime: Long = Date().time,
-)
