--- app/src/main/java/com/example/roomer/domain/model/RecommendedRoommate.kt
@@ -1,8 +1,0 @@
-package com.example.roomer.domain.model
-
-data class RecommendedRoommate(
-    val id: Int,
-    val name: String,
-    val rating: Double,
-    val imagePath: String = "",
-)
