--- app/src/main/java/com/example/roomer/domain/model/SearchUserResult.kt
@@ -1,9 +1,0 @@
-package com.example.roomer.domain.model
-
-data class SearchUserResult(
-    val name: String,
-    val location: String = "",
-    val status: String = "Occasionally",
-    val rate: String,
-    val avatarPath: String = "",
-)
