+++ app/src/main/java/com/example/roomer/domain/usecase/InterestsUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.roomer.domain.usecase
 
-import com.example.roomer.data.repository.RoomerRepositoryInterface
+import com.example.roomer.data.repository.AuthRepositoryInterface
 import com.example.roomer.domain.model.signup.interests.InterestModel
 import com.example.roomer.utils.ConstUseCase
 import com.example.roomer.utils.Resource
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
 class InterestsUseCase(
-    private val repository: RoomerRepositoryInterface
+    private val repository: AuthRepositoryInterface
 ) {
     fun loadInterests(): Flow<Resource<List<InterestModel>>> = flow {
         try {
