+++ app/src/main/java/com/example/roomer/domain/usecase/signup/SignUpOneUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.roomer.domain.usecase.signup
 
-import com.example.roomer.data.repository.RoomerRepositoryInterface
+import com.example.roomer.data.repository.AuthRepositoryInterface
 import com.example.roomer.utils.ConstUseCase
 import com.example.roomer.utils.Resource
 import java.io.IOException
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
 class SignUpOneUseCase(
-    private val repository: RoomerRepositoryInterface
+    private val repository: AuthRepositoryInterface
 ) {
 
     operator fun invoke(
