+++ app/src/main/java/com/example/roomer/domain/usecase/signup/SignUpTwoUseCase.kt
@@ -1,7 +1,7 @@
 package com.example.roomer.domain.usecase.signup
 
 import android.graphics.Bitmap
-import com.example.roomer.data.repository.RoomerRepositoryInterface
+import com.example.roomer.data.repository.AuthRepositoryInterface
 import com.example.roomer.utils.ConstUseCase
 import com.example.roomer.utils.Resource
 import java.io.IOException
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 
 class SignUpTwoUseCase(
-    private val repository: RoomerRepositoryInterface
+    private val repository: AuthRepositoryInterface
 ) {
     operator fun invoke(
         token: String,
