+++ app/src/main/java/com/example/roomer/presentation/screens/entrance/signup/screen_three/SignUpScreenThreeViewModel.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.roomer.data.repository.RoomerRepositoryInterface
+import com.example.roomer.data.repository.AuthRepositoryInterface
 import com.example.roomer.domain.usecase.signup.SignUpThreeUseCase
 import com.example.roomer.utils.Resource
 import com.example.roomer.utils.SpManager
@@ -16,7 +16,7 @@ import kotlinx.coroutines.launch
 @HiltViewModel
 class SignUpScreenThreeViewModel @Inject constructor(
     application: Application,
-    roomerRepository: RoomerRepositoryInterface
+    roomerRepository: AuthRepositoryInterface
 ) : AndroidViewModel(application) {
     private val _state = mutableStateOf(SignUpThreeState())
