--- app/src/main/java/com/example/roomer/utils/navigation.kt
@@ -1,15 +1,0 @@
-package com.example.roomer.utils
-
-/*@Composable
-fun NavbarHostContainer(navController: NavHostController) {
-    NavHost(
-        navController = navController,
-        startDestination = NavbarItem.Home.name,
-    ) {
-        NavbarItem.values().forEach {
-            val navbarItem = it
-            navbarItem.navHostController = navController
-            composable(route = it.name) { navbarItem.composeViewFunction.invoke() }
-        }
-    }
-}*/
